+++ arm9/source/romBrowser/viewModels/RomBrowserAppBarViewModel.h
@@ -53,6 +53,11 @@ public:
         _romBrowserController->ShowFavorites();
     }
 
+    void ShowStatistics()
+    {
+        _romBrowserController->ShowStatistics();
+    }
+
     void RequestDeleteSelected()
     {
         _romBrowserController->RequestDeleteSelected();